# app/routes/extension.py
"""Canonical Core Extension Host management API."""

import typing

import fastapi
import pydantic

from app.business.extension import (
  EXTENSION_HOST,
  EXTENSION_MANAGEMENT_CAPABILITY,
  InstalledExtension,
)
from app.business.peer import PeerHTTPInbound, PeerManager
from app.business.extension.errors import (
  ExtensionAcquisitionError,
  ExtensionCompatibilityError,
  ExtensionEntryPointError,
  ExtensionHostError,
  ExtensionNotInstalledError,
  ExtensionRegistryError,
  ExtensionRuntimeError,
  ExtensionStateConflictError,
)
from app.schemas.extension import ExtensionManagementCommand
from app.schemas.extension.main import EnableExtensionCommand, DisableExtensionCommand
from app.schemas.peer import PeerRef

from .validation import request_input


ROUTER = fastapi.APIRouter(tags=["extension"])
PEER_INBOUND = PeerHTTPInbound(
  capability=EXTENSION_MANAGEMENT_CAPABILITY,
  method="POST",
  path="/extension-management",
)


def _coordinate(namespace: str, name: str) -> str:
  return f"{namespace}/{name}"


def _raise_http_error(error: ExtensionHostError) -> typing.NoReturn:
  if isinstance(error, ExtensionNotInstalledError):
    status_code = fastapi.status.HTTP_404_NOT_FOUND
  elif isinstance(error, ExtensionRegistryError):
    status_code = fastapi.status.HTTP_502_BAD_GATEWAY
  elif isinstance(
    error,
    (
      ExtensionAcquisitionError,
      ExtensionCompatibilityError,
      ExtensionEntryPointError,
      ExtensionRuntimeError,
      ExtensionStateConflictError,
    ),
  ):
    status_code = fastapi.status.HTTP_409_CONFLICT
  else:
    status_code = fastapi.status.HTTP_500_INTERNAL_SERVER_ERROR
  raise fastapi.HTTPException(status_code=status_code, detail=str(error)) from error


@ROUTER.get("/extensions")
def list_extensions(
  limit: int | None = fastapi.Query(None, gt=0), cursor: str | None = None
) -> dict[str, typing.Any]:
  rows = sorted(
    (row for row in EXTENSION_HOST.list() if cursor is None or row.name > cursor),
    key=lambda row: row.name,
  )
  more = limit is not None and len(rows) > limit
  rows = rows[:limit]
  return {"extensions": rows, "next_cursor": rows[-1].name if more else None}


@ROUTER.get("/extensions/{namespace}/{name}")
def get_extension(namespace: str, name: str) -> InstalledExtension:
  try:
    return EXTENSION_HOST.get(_coordinate(namespace, name))
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.post("/extensions/{namespace}/{name}")
def install_extension(
  namespace: str,
  name: str,
  version: str = fastapi.Query(...),
) -> InstalledExtension:
  """Install one exact published Extension Release with no enabled peers."""
  try:
    return EXTENSION_HOST.install(_coordinate(namespace, name), version)
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.delete("/extensions/{namespace}/{name}", status_code=204)
def uninstall_extension(namespace: str, name: str) -> fastapi.Response:
  try:
    EXTENSION_HOST.uninstall(_coordinate(namespace, name))
  except ExtensionHostError as error:
    _raise_http_error(error)
  return fastapi.Response(status_code=fastapi.status.HTTP_204_NO_CONTENT)


@ROUTER.put("/extensions/{namespace}/{name}/config")
def update_extension_config(
  namespace: str,
  name: str,
  body: dict[str, typing.Any] = fastapi.Body(...),
) -> InstalledExtension:
  try:
    with request_input():
      return EXTENSION_HOST.update_config(_coordinate(namespace, name), body)
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.patch("/extensions/{namespace}/{name}/config")
def patch_extension_config(
  namespace: str, name: str, body: dict[str, typing.Any] = fastapi.Body(...)
) -> InstalledExtension:
  try:
    with request_input():
      return EXTENSION_HOST.patch_config(_coordinate(namespace, name), body)
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.post("/extensions/{namespace}/{name}/enable")
async def enable_extension(
  namespace: str, name: str, route_to_peer: PeerRef | None = None
) -> InstalledExtension:
  try:
    return await EXTENSION_HOST.manage(
      EnableExtensionCommand(action="enable", extension=_coordinate(namespace, name)),
      route_to_peer=route_to_peer or PeerManager.get_current_peer_ref(),
    )
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.post("/extensions/{namespace}/{name}/disable")
async def disable_extension(
  namespace: str, name: str, route_to_peer: PeerRef | None = None
) -> InstalledExtension:
  try:
    return await EXTENSION_HOST.manage(
      DisableExtensionCommand(action="disable", extension=_coordinate(namespace, name)),
      route_to_peer=route_to_peer or PeerManager.get_current_peer_ref(),
    )
  except ExtensionHostError as error:
    _raise_http_error(error)


@ROUTER.post("/extension-management", include_in_schema=False)
async def manage_extension(body: ExtensionManagementCommand) -> InstalledExtension:
  """Execute the fixed Peer-local Extension management capability."""
  try:
    return await EXTENSION_HOST.manage_local(body)
  except pydantic.ValidationError as error:
    raise fastapi.HTTPException(
      status_code=fastapi.status.HTTP_422_UNPROCESSABLE_CONTENT,
      detail=error.errors(include_url=False, include_context=False),
    ) from error
  except ExtensionHostError as error:
    _raise_http_error(error)